faq-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
openai==1.35.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
gunicorn==20.1.0
